backend/pyproject.toml
[project]
name = "spx-exam"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "duckdb>=1.3.2",
    "fastapi>=0.116.1",
    "gunicorn>=23.0.0",
    "ipykernel>=6.30.1",
    "ipywidgets>=8.1.7",
    "nbformat>=5.10.4",
    "openai-agents[litellm]>=0.2.11",
    "pandas>=2.3.2",
    "plotly>=6.3.0",
    "polars>=1.33.0",
    "psycopg2-binary>=2.9.10",
    "requests>=2.32.5",
    "sqlalchemy>=2.0.43",
    "uvicorn>=0.35.0",
]
